requirements.txt
flask
python-dotenv
flask-wtf
email_validator
flask-sqlalchemy
flask-migrate
flask-login
flask-mail
pyjwt
flask-moment
flask-babel
black
langdetect
requests
elasticsearch
